Revert "Update 3.3.5 (#6)"
This reverts commit 273884709e423f648eeb6be97ab03b16b011171e.

+++ src/server/game/Server/Packets/MiscPackets.cpp
@@ -204,7 +204,7 @@ void WorldPackets::Misc::WorldTeleport::Read()
 
 WorldPacket const* WorldPackets::Misc::CorpseReclaimDelay::Write()
 {
-    _worldPacket << uint32(Remaining);
+    _worldPacket << Remaining;
 
     return &_worldPacket;
 }
